requirements.txt
asgiref==3.2.3
Babel==2.8.0
backcall==0.1.0
certifi==2019.11.28
decorator==4.4.1
Django==3.0.2
django-appconf==1.0.3
django-extensions==2.2.6
django-imagekit==2.0.4
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.16.0
parso==0.6.0
pexpect==4.8.0
phonenumbers==8.11.2
pickleshare==0.7.5
prompt-toolkit==3.0.3
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
pytz==2019.3
sentry-sdk==0.14.1
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
